feat(change webpack config): 修改webpack配置
