+++ package.json
@@ -29,7 +29,6 @@
     "start": "cross-env NODE_ENV=development react-app-rewired start",
     "dev": "cross-env NODE_ENV=development webpack serve",
     "build": "cross-env NODE_ENV=production webpack",
-    "test": "cross-env NODE_ENV=test react-app-rewired test",
     "prepare": "husky install",
     "release": "standard-version"
   },
@@ -70,12 +69,15 @@
     "html-webpack-plugin": "^5.5.3",
     "husky": "^8.0.3",
     "interpolate-html-plugin": "^4.0.0",
+    "less-loader": "^11.1.3",
     "mini-css-extract-plugin": "^2.7.6",
     "react-app-rewired": "^2.2.1",
+    "sass-loader": "^13.3.2",
     "standard-version": "^9.5.0",
     "style-loader": "^3.3.3",
     "webpack": "^5.88.2",
     "webpack-cli": "^5.1.4",
-    "webpack-dev-server": "^4.15.1"
+    "webpack-dev-server": "^4.15.1",
+    "webpack-merge": "^5.10.0"
   }
 }
