+++ webpack.common.js
@@ -1,0 +1,36 @@
+const path = require('path')
+const HtmlWebpackPlugin = require('html-webpack-plugin')
+
+module.exports = {
+    entry: './src/index.tsx',
+    output: {
+        filename: 'main.js',
+        path: path.resolve(__dirname, 'dist'),
+        clean: true
+    },
+    module: {
+        rules: [
+            {
+                test: /.(ts|tsx)$/,
+                exclude: /(node_modules)/,
+                use: {
+                    loader: 'babel-loader',
+                    options: {
+                        presets: ['@babel/preset-env', '@babel/preset-react', '@babel/preset-typescript']
+                    }
+                }
+            }
+        ]
+    },
+    resolve: {
+        extensions: ['.js', '.tsx', '.ts', '.jsx'],
+        alias: {
+            '@': path.resolve(__dirname, 'src')
+        }
+    },
+    plugins: [
+        new HtmlWebpackPlugin({
+            template: path.resolve(__dirname, 'public/index.html')
+        })
+    ]
+}
